refactor: centralize V4L2 device management logic
